+++ website/app/helpers/hotdocs_helper.rb
@@ -71,4 +71,8 @@ module HotdocsHelper
       }
     ]
   end
+
+  def fetcher_host
+    "http://127.0.0.1:3000"
+  end
 end