+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Fetcher helper to fetch fresh content from a static page
 - Prevent possible (implicit) styles overwrite
 
 ## [0.2.0]